requirements.txt
treelib
pyserial
requests
ws4py
websocket-client
pytest
pytest-cov
virtualenv
esptool
pytest-timeout
timeout_decorator
